removed uneeded files
